Further changes to make Deepspeed work

+++ finetuning/models.py
@@ -618,15 +618,15 @@ class PairwiseWithOriginalDataJointTraining(L.LightningModule):
 
             if i == 0:  # this is the pairwise data
                 X1, X2, Y = (
-                    dl_batch["seq1"],
-                    dl_batch["seq2"],
-                    dl_batch["z_diff"].float(),
+                    dl_batch["seq1"].half(),
+                    dl_batch["seq2"].half(),
+                    dl_batch["z_diff"].half(),
                 )
                 mse_loss = self.get_mse_loss(X1, X2, Y)
                 self.log("train/pairwise_mse_loss", mse_loss)
                 loss += mse_loss
             elif i == 1:  # this is the original human training data
-                X, Y = dl_batch["seq"], dl_batch["y"].float()
+                X, Y = dl_batch["seq"].half(), dl_batch["y"].half()
                 Y_hat = self(
                     X, return_base_predictions=True, base_predictions_head="human"
                 )
@@ -634,7 +634,7 @@ class PairwiseWithOriginalDataJointTraining(L.LightningModule):
                 self.log("train/human_poisson_loss", poisson_loss)
                 loss += poisson_loss
             elif i == 2:  # this is the original mouse training data
-                X, Y = dl_batch["seq"], dl_batch["y"].float()
+                X, Y = dl_batch["seq"].half(), dl_batch["y"].half()
                 Y_hat = self(
                     X, return_base_predictions=True, base_predictions_head="mouse"
                 )
@@ -651,12 +651,16 @@ class PairwiseWithOriginalDataJointTraining(L.LightningModule):
 
     def validation_step(self, batch, batch_idx, dataloader_idx=0):
         if dataloader_idx == 0:
-            X1, X2, Y = batch["seq1"], batch["seq2"], batch["z_diff"].float()
+            X1, X2, Y = (
+                batch["seq1"].half(),
+                batch["seq2"].half(),
+                batch["z_diff"].half(),
+            )
             mse_loss = self.get_mse_loss(X1, X2, Y)
             self.log("val/pairwise_mse_loss", mse_loss)
 
         elif dataloader_idx == 1:
-            X, Y = batch["seq"], batch["y"].float()
+            X, Y = batch["seq"].half(), batch["y"].half()
             Y_hat = self(X, return_base_predictions=True, base_predictions_head="human")
             poisson_loss = self.poisson_loss(Y_hat, Y)
             self.log("val/human_poisson_loss", poisson_loss)
@@ -667,7 +671,7 @@ class PairwiseWithOriginalDataJointTraining(L.LightningModule):
             self.human_metrics["mse"].update(Y_hat, Y)
 
         elif dataloader_idx == 2:
-            X, Y = batch["seq"], batch["y"].float()
+            X, Y = batch["seq"].half(), batch["y"].half()
             Y_hat = self(X, return_base_predictions=True, base_predictions_head="mouse")
             poisson_loss = self.poisson_loss(Y_hat, Y)
             self.log("val/mouse_poisson_loss", poisson_loss)
